+++ examples/custom-echo-soul/main.py
@@ -1,13 +1,11 @@
 import asyncio
 from typing import Any
 
-from kosong.message import ContentPart, TextPart
-
 from kimi_cli.llm import ALL_MODEL_CAPABILITIES, ModelCapability
 from kimi_cli.soul import StatusSnapshot, wire_send
 from kimi_cli.ui.shell import Shell
 from kimi_cli.utils.slashcmd import SlashCommand
-from kimi_cli.wire.message import StepBegin
+from kimi_cli.wire.types import ContentPart, StepBegin, TextPart
 
 
 class EchoSoul: